requirements.txt
apispec==6.4.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-DotEnv==0.1.2
flask-smorest==0.42.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.20.2
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
validators==0.22.0
webargs==8.4.0
Werkzeug==3.0.1
